requirements.txt
Flask
pydicom
Pillow
numpy